Removed remained of warnings.  Note, in CPU.cpp a couple of warning types had to be suppressed because its a specialized situation.

+++ src/MouseInterface.cpp
@@ -478,7 +478,7 @@ void CMouseInterface::OnMouseEvent()
 
 	BOOL bBtn0 = m_bButtons[0];
 	BOOL bBtn1 = m_bButtons[1];
-	if ( m_nX != m_iX || m_nY != m_iY )
+	if ( (UINT) m_nX != m_iX || (UINT) m_nY != m_iY )
 		byState |= 0x22;				// X/Y moved since last READMOUSE | Movement interrupt
 	if ( m_bBtn0 != bBtn0 || m_bBtn1 != bBtn1 )
 		byState |= 0x04;				// Button 0/1 interrupt
